+++ internal/plans/internal/planproto/planfile.pb.go
@@ -89,7 +89,6 @@ const (
 	Action_DELETE_THEN_CREATE Action = 6
 	Action_CREATE_THEN_DELETE Action = 7
 	Action_FORGET             Action = 8
-	Action_OPEN               Action = 9
 	Action_FORGET_THEN_CREATE Action = 10
 )
 
@@ -104,7 +103,6 @@ var (
 		6:  "DELETE_THEN_CREATE",
 		7:  "CREATE_THEN_DELETE",
 		8:  "FORGET",
-		9:  "OPEN",
 		10: "FORGET_THEN_CREATE",
 	}
 	Action_value = map[string]int32{
@@ -116,7 +114,6 @@ var (
 		"DELETE_THEN_CREATE": 6,
 		"CREATE_THEN_DELETE": 7,
 		"FORGET":             8,
-		"OPEN":               9,
 		"FORGET_THEN_CREATE": 10,
 	}
 )
@@ -1467,7 +1464,7 @@ const file_planfile_proto_rawDesc = "" +
 	"\n" +
 	"\x06NORMAL\x10\x00\x12\v\n" +
 	"\aDESTROY\x10\x01\x12\x10\n" +
-	"\fREFRESH_ONLY\x10\x02*\x9e\x01\n" +
+	"\fREFRESH_ONLY\x10\x02*\x9a\x01\n" +
 	"\x06Action\x12\b\n" +
 	"\x04NOOP\x10\x00\x12\n" +
 	"\n" +
@@ -1480,10 +1477,9 @@ const file_planfile_proto_rawDesc = "" +
 	"\x12DELETE_THEN_CREATE\x10\x06\x12\x16\n" +
 	"\x12CREATE_THEN_DELETE\x10\a\x12\n" +
 	"\n" +
-	"\x06FORGET\x10\b\x12\b\n" +
-	"\x04OPEN\x10\t\x12\x16\n" +
+	"\x06FORGET\x10\b\x12\x16\n" +
 	"\x12FORGET_THEN_CREATE\x10\n" +
-	"*\xc9\x04\n" +
+	"\"\x04\b\t\x10\t*\xc9\x04\n" +
 	"\x1cResourceInstanceActionReason\x12\b\n" +
 	"\x04NONE\x10\x00\x12\x1b\n" +
 	"\x17REPLACE_BECAUSE_TAINTED\x10\x01\x12\x16\n" +