+++ internal/tofu/context_plan2_test.go
@@ -2685,94 +2685,6 @@ func TestContext2Plan_refreshOnlyMode(t *testing.T) {
 	}
 }
 
-func TestContext2Plan_refreshOnlyMode_ephemeral(t *testing.T) {
-	addr := mustResourceInstanceAddr("ephemeral.test_object.a")
-
-	// The configuration, the prior state, and the refresh result intentionally
-	// have different values for "test_string" so we can observe that the
-	// refresh took effect but the configuration change wasn't considered.
-	m := testModuleInline(t, map[string]string{
-		"main.tf": `
-			ephemeral "test_object" "a" {
-				arg = "after"
-			}
-		`,
-	})
-	state := states.NewState()
-
-	p := simpleMockProvider()
-	p.GetProviderSchemaResponse = &providers.GetProviderSchemaResponse{
-		Provider: providers.Schema{Block: simpleTestSchema()},
-		EphemeralResources: map[string]providers.Schema{
-			"test_object": {
-				Block: &configschema.Block{
-					Attributes: map[string]*configschema.Attribute{
-						"arg": {Type: cty.String, Optional: true},
-					},
-				},
-			},
-		},
-	}
-	p.OpenEphemeralResourceFn = func(req providers.OpenEphemeralResourceRequest) providers.OpenEphemeralResourceResponse {
-		newVal, err := cty.Transform(req.Config, func(path cty.Path, v cty.Value) (cty.Value, error) {
-			if len(path) == 1 && path[0] == (cty.GetAttrStep{Name: "arg"}) {
-				return cty.StringVal("current"), nil
-			}
-			return v, nil
-		})
-		if err != nil {
-			// shouldn't get here
-			t.Fatalf("OpenResourceFn transform failed")
-			return providers.OpenEphemeralResourceResponse{}
-		}
-		return providers.OpenEphemeralResourceResponse{
-			Result: newVal,
-		}
-	}
-
-	ctx := testContext2(t, &ContextOpts{
-		Plugins: plugins.NewLibrary(map[addrs.Provider]providers.Factory{
-			addrs.NewDefaultProvider("test"): testProviderFuncFixed(p),
-		}, nil),
-	})
-
-	plan, diags := ctx.Plan(context.Background(), m, state, &PlanOpts{
-		Mode: plans.RefreshOnlyMode,
-	})
-	if diags.HasErrors() {
-		t.Fatalf("unexpected errors\n%s", diags.Err().Error())
-	}
-
-	if !p.OpenEphemeralResourceCalled {
-		t.Errorf("Provider's OpenEphemeralResource wasn't called; should've been")
-	}
-
-	if got, want := len(plan.Changes.Resources), 1; got != want {
-		t.Fatalf("expected to have exactly %d resource but got %d", want, got)
-	}
-	if gotResAddr := plan.Changes.Resources[0].Addr; !gotResAddr.Equal(addr) {
-		t.Errorf("plan contains one resource and that's NOT an ephemeral as expected; instead, got %s", gotResAddr)
-	}
-	if got, want := len(plan.Changes.ActionableResources()), 0; got != want {
-		t.Errorf(
-			"changes.ActionableResources() returned more than %d resources, meaning that didn't exclude ephemeral resources. Instead returned %d\nChanges:\n%s",
-			want,
-			got,
-			spew.Sdump(plan.Changes.Resources),
-		)
-	}
-
-	if instState := plan.PlannedState.ResourceInstance(addr); instState == nil {
-		t.Errorf("%s has no planned state, but it should have since it's needed to build the apply graph correctly", addr)
-	} else {
-		want := `{"arg":"current"}`
-		got := string(instState.Current.AttrsJSON)
-		if diff := cmp.Diff(want, got); diff != "" {
-			t.Fatalf("unexpected attributes for the planned ephemeral:\n%s", diff)
-		}
-	}
-}
-
 func TestContext2Plan_refreshOnlyMode_deposed(t *testing.T) {
 	addr := mustResourceInstanceAddr("test_object.a")
 	deposedKey := states.DeposedKey("byebye")
@@ -8872,16 +8784,17 @@ func TestContext2Plan_insufficient_block(t *testing.T) {
 }
 
 // Ensure that running plan on a configuration with ephemeral resources,
-// the generated plan contains the expected changes
-func TestContext2Plan_ephemeralResourceChangesGenerated(t *testing.T) {
+// the plan contains no changes for the ephemerals.
+// This test has been repurposed during #3799.
+func TestContext2Plan_noEphemeralResourceChangesGenerated(t *testing.T) {
 	m := testModuleInline(t, map[string]string{
 		"main.tf": `
 ephemeral "test_ephemeral_resource" "a" {
 }
 `,
 	})
-	testProvider := testProvider("test")
-	testProvider.OpenEphemeralResourceResponse = &providers.OpenEphemeralResourceResponse{
+	p := testProvider("test")
+	p.OpenEphemeralResourceResponse = &providers.OpenEphemeralResourceResponse{
 		Result: cty.ObjectVal(map[string]cty.Value{
 			"id":     cty.StringVal("id val"),
 			"secret": cty.StringVal("val"),
@@ -8893,49 +8806,24 @@ ephemeral "test_ephemeral_resource" "a" {
 
 	ctx := testContext2(t, &ContextOpts{
 		Plugins: plugins.NewLibrary(map[addrs.Provider]providers.Factory{
-			addrs.NewDefaultProvider("test"): testProviderFuncFixed(testProvider),
+			addrs.NewDefaultProvider("test"): testProviderFuncFixed(p),
 		}, nil),
 	})
 
-	plan, diags := ctx.Plan(context.Background(), m, state, DefaultPlanOpts)
-	if diags.HasErrors() {
-		t.Fatalf("unexpected plan error: %s", diags)
-	}
-	if plan.Changes == nil {
-		t.Fatalf("expected to have some changes but got none")
-	}
-	if got, want := len(plan.Changes.Resources), 1; got != want {
-		t.Fatalf("expected to have %d changes but got %d", want, got)
-	}
-	got := plan.Changes.Resources[0]
-	addr := mustResourceInstanceAddr("ephemeral.test_ephemeral_resource.a")
-	schema := testProvider.ProviderSchema().EphemeralTypes[addr.Resource.Resource.Type]
-	objTy := schema.ImpliedType()
-	priorVal := cty.NullVal(objTy)
-	beforeVal, err := plans.NewDynamicValue(priorVal, objTy)
-	if err != nil {
-		t.Fatalf("unexpected error creating before val: %s", err)
-	}
-	afterVal, err := plans.NewDynamicValue(cty.ObjectVal(map[string]cty.Value{
-		"id":     cty.StringVal("id val"),
-		"secret": cty.StringVal("val"),
-		"input":  cty.NullVal(cty.String),
-	}), objTy)
-	if err != nil {
-		t.Fatalf("unexpected error creating after val: %s", err)
-	}
-	want := &plans.ResourceInstanceChangeSrc{
-		Addr:         addr,
-		PrevRunAddr:  addr,
-		ProviderAddr: mustProviderConfig(`provider["registry.opentofu.org/hashicorp/test"]`),
-		ChangeSrc: plans.ChangeSrc{
-			Action: plans.Open,
-			Before: beforeVal,
-			After:  afterVal,
-		},
-	}
-	if diff := cmp.Diff(want, got); diff != "" {
-		t.Fatalf("unexpected diff in the ephemeral resource recorded change:\n%s", diff)
+	for _, mode := range []plans.Mode{plans.NormalMode, plans.RefreshOnlyMode} {
+		t.Run(mode.String(), func(t *testing.T) {
+			plan, diags := ctx.Plan(context.Background(), m, state, &PlanOpts{Mode: mode})
+			if diags.HasErrors() {
+				t.Fatalf("unexpected plan error: %s", diags)
+			}
+			if !p.OpenEphemeralResourceCalled {
+				t.Errorf("Provider's OpenEphemeralResource wasn't called; should've been")
+			}
+			defer func() { p.OpenEphemeralResourceCalled = false }()
+			if got, want := len(plan.Changes.Resources), 0; got != want {
+				t.Fatalf("expected to have %d changes but got %d", want, got)
+			}
+		})
 	}
 }