+++ internal/tofu/evaluate_test.go
@@ -648,6 +648,7 @@ func TestEvaluatorGetResource_Ephemeral(t *testing.T) {
 		Name: "foo",
 		Config: configs.SynthBody("", map[string]cty.Value{
 			"secret_name": cty.StringVal("foo"),
+			"name":        cty.StringVal("bar"),
 		}),
 		Provider: mustProviderConfig(`provider["registry.opentofu.org/hashicorp/test"]`).Provider,
 	}
@@ -657,6 +658,10 @@ func TestEvaluatorGetResource_Ephemeral(t *testing.T) {
 				"id": {
 					Type:     cty.String,
 					Computed: true,
+				},
+				"name": {
+					Type:     cty.String,
+					Required: true,
 				},
 				"value": {
 					Type:     cty.String,
@@ -681,56 +686,20 @@ func TestEvaluatorGetResource_Ephemeral(t *testing.T) {
 		},
 	}
 	tests := map[string]struct {
-		changes *plans.ChangesSync
-		state   *states.SyncState
-		want    cty.Value
+		state *states.SyncState
+		want  cty.Value
 	}{
-		"no changes and no state": {
-			plans.NewChanges().SyncWrapper(),
+		"no state": {
 			states.NewState().SyncWrapper(),
 			cty.DynamicVal.Mark(marks.Ephemeral),
 		},
-		"with state and planned changes": {
-			plans.BuildChanges(func(sync *plans.ChangesSync) {
-				sync.AppendResourceInstanceChange(
-					&plans.ResourceInstanceChangeSrc{
-						Addr:        rc.Addr().Instance(addrs.NoKey).Absolute(addrs.RootModuleInstance),
-						PrevRunAddr: rc.Addr().Instance(addrs.NoKey).Absolute(addrs.RootModuleInstance),
-						DeposedKey:  states.NotDeposed,
-						ProviderAddr: addrs.AbsProviderConfig{
-							Provider: rc.Provider,
-							Module:   addrs.RootModule,
-						},
-						ChangeSrc: plans.ChangeSrc{
-							After: encodeDynamicValue(t, cty.ObjectVal(map[string]cty.Value{
-								"id":    cty.StringVal("foo"),
-								"value": cty.StringVal("tacos"),
-								"nesting_map": cty.SetVal([]cty.Value{
-									cty.ObjectVal(map[string]cty.Value{
-										"foo": cty.StringVal("test"),
-									}),
-								}),
-							})),
-							AfterValMarks: []cty.PathValueMarks{
-								{
-									Path: cty.GetAttrPath("nesting_map").Index(cty.ObjectVal(map[string]cty.Value{"foo": cty.StringVal("test")})).GetAttr("foo"),
-									Marks: map[interface{}]struct{}{
-										// added the ephemeral mark here to validate that it is removed and the
-										// sensitive one is added based on the schema
-										marks.Ephemeral: {},
-									},
-								},
-							},
-						},
-					},
-				)
-			}).SyncWrapper(),
+		"with state": {
 			states.BuildState(func(state *states.SyncState) {
 				state.SetResourceInstanceCurrent(
 					rc.Addr().Instance(addrs.NoKey).Absolute(addrs.RootModuleInstance),
 					&states.ResourceInstanceObjectSrc{
 						Status:    states.ObjectPlanned,
-						AttrsJSON: []byte(`{"id":"foo", "val":"tacos"}`),
+						AttrsJSON: []byte(`{"id": "foo", "name": "bar", "value": "tacos", "nesting_map": [{"foo":"test"}]}`),
 					},
 					addrs.AbsProviderConfig{
 						Provider: rc.Provider,
@@ -742,6 +711,7 @@ func TestEvaluatorGetResource_Ephemeral(t *testing.T) {
 			cty.ObjectVal(map[string]cty.Value{
 				"id":    cty.StringVal("foo"),
 				"value": cty.StringVal("tacos"),
+				"name":  cty.StringVal("bar"),
 				"nesting_map": cty.SetVal([]cty.Value{
 					cty.ObjectVal(map[string]cty.Value{
 						// expected to have this attribute marked as sensitive but not as ephemeral
@@ -751,14 +721,37 @@ func TestEvaluatorGetResource_Ephemeral(t *testing.T) {
 				}),
 			}).Mark(marks.Ephemeral),
 		},
-		"with object ready state and no changes": {
-			plans.BuildChanges(func(sync *plans.ChangesSync) {}).SyncWrapper(),
+		"with defered state": {
+			states.BuildState(func(state *states.SyncState) {
+				state.SetResourceInstanceCurrent(
+					rc.Addr().Instance(addrs.NoKey).Absolute(addrs.RootModuleInstance),
+					&states.ResourceInstanceObjectSrc{
+						Status: states.ObjectPlanned,
+						// NOTE: during decoding, null values will be converted to unknown
+						AttrsJSON: []byte(`{"id":null,"name":"bar","nesting_map":null,"value":null}`),
+						Deferred:  true,
+					},
+					addrs.AbsProviderConfig{
+						Provider: rc.Provider,
+						Module:   addrs.RootModule,
+					},
+					addrs.NoKey,
+				)
+			}).SyncWrapper(),
+			cty.ObjectVal(map[string]cty.Value{
+				"id":          cty.UnknownVal(cty.String),
+				"value":       cty.UnknownVal(cty.String),
+				"name":        cty.StringVal("bar"),
+				"nesting_map": cty.NullVal(cty.Set(cty.Object(map[string]cty.Type{"foo": cty.String}))),
+			}).Mark(marks.Ephemeral),
+		},
+		"with object ready state": {
 			states.BuildState(func(state *states.SyncState) {
 				state.SetResourceInstanceCurrent(
 					rc.Addr().Instance(addrs.NoKey).Absolute(addrs.RootModuleInstance),
 					&states.ResourceInstanceObjectSrc{
 						Status:    states.ObjectReady,
-						AttrsJSON: []byte(`{"id":"foo", "value":"tacos", "nesting_map": [{"foo": "test"}]}`),
+						AttrsJSON: []byte(`{"id":"foo", "name": "bar", "value":"tacos", "nesting_map": [{"foo": "test"}]}`),
 					},
 					addrs.AbsProviderConfig{
 						Provider: rc.Provider,
@@ -770,6 +763,7 @@ func TestEvaluatorGetResource_Ephemeral(t *testing.T) {
 			cty.ObjectVal(map[string]cty.Value{
 				"id":    cty.StringVal("foo"),
 				"value": cty.StringVal("tacos"),
+				"name":  cty.StringVal("bar"),
 				"nesting_map": cty.SetVal([]cty.Value{
 					cty.ObjectVal(map[string]cty.Value{
 						// expected to have this attribute marked as sensitive but not as ephemeral
@@ -784,14 +778,13 @@ func TestEvaluatorGetResource_Ephemeral(t *testing.T) {
 		t.Run(name, func(t *testing.T) {
 			// having these here for easier reference in the test body
 			state := tt.state
-			changes := tt.changes
 			want := tt.want
 
 			evaluator := &Evaluator{
 				Meta: &ContextMeta{
 					Env: "foo",
 				},
-				Changes: changes,
+				Changes: plans.NewChanges().SyncWrapper(),
 				Config: &configs.Config{
 					Module: &configs.Module{
 						EphemeralResources: map[string]*configs.Resource{